MT#56772 AmSdp: covert sessId / sessV to llu
Use `unsigned long long` for SDP session id and version
instead of unsigned int.

Refactor all usage of them accordingly.

Additionally intrdouce new utils functions for conversion:
- `ulonglong2str()` - converts `unsigned long long` to `string`
- `str2ull()` - converts `string` to `unsigned long long`

+++ core/AmSdp.h
@@ -81,8 +81,8 @@ struct SdpConnection
 struct SdpOrigin
 {
   string user;
-  unsigned int sessId;
-  unsigned int sessV;
+  unsigned long long sessId;
+  unsigned long long sessV;
   SdpConnection conn;
 
   SdpOrigin() : user(), conn() {}
